refactor pt 2
