+++ src/main/kotlin/frc/team449/commands/light/PickupBlink.kt
@@ -1,10 +1,10 @@
-package frc.team449.robot2024.commands.light
+package frc.team449.commands.light
 
 import edu.wpi.first.wpilibj2.command.Command
 import edu.wpi.first.wpilibj2.command.InstantCommand
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup
 import edu.wpi.first.wpilibj2.command.WaitCommand
-import frc.team449.system.light.Light
+import frc.team449.subsystems.light.Light
 
 /** Description: Blink a certain color 5 times */
 class PickupBlink {
