+++ src/main/kotlin/frc/team449/commands/light/Rainbow.kt
@@ -1,9 +1,9 @@
-package frc.team449.robot2024.commands.light
+package frc.team449.commands.light
 
 import edu.wpi.first.math.MathUtil
 import edu.wpi.first.wpilibj2.command.Command
-import frc.team449.robot2024.constants.subsystem.LightConstants
-import frc.team449.system.light.Light
+import frc.team449.subsystems.light.LightConstants
+import frc.team449.subsystems.light.Light
 
 class Rainbow(
   private val led: Light
