+++ src/main/kotlin/frc/team449/commands/light/SolidColor.kt
@@ -1,7 +1,7 @@
-package frc.team449.robot2024.commands.light
+package frc.team449.commands.light
 
 import edu.wpi.first.wpilibj2.command.Command
-import frc.team449.system.light.Light
+import frc.team449.subsystems.light.Light
 
 /** Description: Have a solid color  */
 class SolidColor(
